address code smells in sonar
